+++ quanmiaolightapp-20240801/core/Models/RunMarketingInformationWritingRequest.cs
@@ -9,9 +9,17 @@ using Tea;
 namespace AlibabaCloud.SDK.QuanMiaoLightApp20240801.Models
 {
     public class RunMarketingInformationWritingRequest : TeaModel {
+        [NameInMap("customLimitation")]
+        [Validation(Required=false)]
+        public string CustomLimitation { get; set; }
+
         [NameInMap("customPrompt")]
         [Validation(Required=false)]
         public string CustomPrompt { get; set; }
+
+        [NameInMap("inputExample")]
+        [Validation(Required=false)]
+        public string InputExample { get; set; }
 
         /// <summary>
         /// <b>Example:</b>
@@ -21,6 +29,10 @@ namespace AlibabaCloud.SDK.QuanMiaoLightApp20240801.Models
         [NameInMap("modelId")]
         [Validation(Required=false)]
         public string ModelId { get; set; }
+
+        [NameInMap("outputExample")]
+        [Validation(Required=false)]
+        public string OutputExample { get; set; }
 
         [NameInMap("sourceMaterial")]
         [Validation(Required=false)]